Adds hardcoded layout for first half of page
