+++ src/index/index.module.css
@@ -1,0 +1,93 @@
+.omVariant {
+  width: 70%;
+  padding-top: 7.2rem;
+  position: relative;
+  margin-bottom: 12rem;
+}
+.omVariant__title {
+  font-weight: bold;
+  margin: 2.4rem 0;
+  font-size: 4.8rem;
+}
+.omVariant__p1 {
+  margin: 2rem 3rem;
+  font-size: 1.5rem;
+}
+
+.omVariant_image1 {
+  position: relative;
+}
+.omVariant_image2 {
+  position: absolute;
+  left: 25rem;
+  top: 32rem;
+}
+.omVariant__us {
+  width: 70%;
+  margin-top: 3.4rem;
+}
+
+.omVariant__us p {
+  margin: 2.4rem 0;
+  font-size: 1.5rem;
+}
+
+.join {
+  text-align: center;
+  max-width: 24rem;
+  margin: 0 auto;
+}
+.join p {
+  margin: 2.4rem 0;
+}
+
+.services {
+  margin: 3rem 0;
+  padding: 0 1.2rem;
+}
+.services h3 {
+  margin-bottom: 2.4rem;
+}
+
+.services_image1 {
+  position: relative;
+  left: -7rem;
+}
+.services article + article {
+  margin-top: 4.8rem;
+  width: 50%;
+}
+.services article:nth-child(2n) {
+  margin-left: 50%;
+}
+.services img + article {
+  margin-top: -10rem;
+}
+
+.blog {
+  text-align: center;
+  margin: -5rem auto 0;
+  padding: 8rem 0;
+  position: relative;
+}
+.blog__text {
+  position: relative;
+
+  z-index: 2;
+}
+.blog_image1 {
+  position: absolute;
+  top: 0;
+  z-index: 1;
+  left: calc(50% + 1rem);
+}
+.blog a {
+  color: var(--color-standard__black);
+  text-decoration: none;
+  margin-top: 2.4rem;
+  display: block;
+}
+.blog img {
+  display: block;
+  margin: 0.6rem auto 0;
+}
